pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "openai>=1.102.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
